Changed exit hook to be more graceful
Exiting comSR620 occurs by sd_hook() which is attched to SIGINT.  But, I
have noticed that the TIC is always confused after doing this.  For
exmaple, the next run of comSR620 will always fail with "not a TIC!",
but then the next run will work, but often, the next run will be plaged
with random zero measurements.  I think the TIC is getting confused.
So, I've updated the hook with a few SR620 commands which stop whatever
measurement is active and stops any auto-measurement that is active.
I also replaced some planned exits that called the hook with direct
exit() calls if I thought it was safe.