wrapper: Add site name variable and defaults
I made an error updating Kenkyuto's boot process last year.  All the
filenames had NU1 in them, instead of Kenkyuto.  This adds a
configuration variable so this can be easily set (and seen) at the
beginning of the wrapper script.  It also implements a set of default
selections and falls back to hostname so that, even if I forget to set
the site name properly in the future, it should try to do the right
thing.